+++ sdk/formrecognizer/Azure.AI.FormRecognizer/src/RecognizeContentOptions.cs
@@ -38,10 +38,17 @@ namespace Azure.AI.FormRecognizer
         public string Language { get; set; }
 
         /// <summary>
+        /// <para>
         /// Custom page numbers for multi-page documents(PDF/TIFF). Input the number of the
         /// pages you want to get OCR result. For a range of pages, use a hyphen.
         /// Separate each page or range with a comma.
+        /// For sample "1-3", {"1", "3", "5-7"}.
+        /// </para>
+        /// <para>
+        /// Although this collection cannot be set, it can be modified.
+        /// See <a href="https://docs.microsoft.com/en-us/dotnet/csharp/programming-guide/classes-and-structs/object-and-collection-initializers#collection-initializers">collection initializer</a>.
+        /// </para>
         /// </summary>
-        public IEnumerable<string> Pages { get; set; }
+        public IList<string> Pages { get; } = new List<string>();
     }
 }