+++ sdk/formrecognizer/Azure.AI.FormRecognizer/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Renamed the model `Appearance` to `TextAppearance`.
 - Renamed the model `Style` to `TextStyle`.
 - Renamed the extensible enum `TextStyle` to `TextStyleName`.
+- Changed object type for property `Pages` under `RecognizeContentOptions` from `IEnumerable` to `IList`.
 
 ## 3.1.0-beta.1 (2020-11-23)